requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
cryptography==45.0.4
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
pyzmq==27.0.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
wcwidth==0.2.13
